Q2/anime-agent/pyproject.toml
[project]
name = "anime-agent"
version = "0.1.0"
description = "An AI agent that helps users explore anime series, genres, terminology, and culture."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
start-agent = "anime.agent:run_anime_agent"

[tool.hatch.build.targets.wheel]
packages = ["src/anime"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
